torchci: extract advisor dispatch into a shared lib + saved CH queries
Pure refactor of the manual "AI Analyze" dispatch endpoint, no behavior
change — split out ahead of the auto-dispatch feature so it can land on
its own.

- New lib/advisor/advisorDispatch.ts: the signal_pattern build + workflow
  dispatch (dispatchAdvisorWorkflow) and helpers (signalKeyForJob,
  isValidSha, the job-status fetchers), extracted verbatim from
  dispatch-advisor.ts so they can be reused by other callers.
- Move the three inline ClickHouse queries into clickhouse_queries/ saved
  queries (advisor_job_status_exact, advisor_job_status_pattern,
  advisor_trunk_shas_with_job), called via queryClickhouseSaved.
- dispatch-advisor.ts becomes a thin handler: same auth + write-permission
  gate, same verdict-based 10-minute dedup, same workflow file and 200/409/
  500 responses — it just delegates the dispatch to the shared lib.

No functional change: the manual button behaves exactly as before.
