+++ torchci/clickhouse_queries/advisor_job_status_exact/params.json
@@ -1,0 +1,8 @@
+{
+  "params": {
+    "repo": "String",
+    "jobName": "String",
+    "shas": "Array(String)"
+  },
+  "tests": []
+}
