+++ torchci/clickhouse_queries/advisor_job_status_pattern/params.json
@@ -1,0 +1,8 @@
+{
+  "params": {
+    "repo": "String",
+    "jobPattern": "String",
+    "shas": "Array(String)"
+  },
+  "tests": []
+}
