+++ torchci/clickhouse_queries/advisor_trunk_shas_with_job/params.json
@@ -1,0 +1,9 @@
+{
+  "params": {
+    "repo": "String",
+    "branch": "String",
+    "jobPattern": "String",
+    "limit": "UInt32"
+  },
+  "tests": []
+}
